+++ web-wallet/src/lib/dusk/components/__tests__/Stepper.spec.js
@@ -57,6 +57,13 @@ describe("Stepper", () => {
     expect(container.firstChild).toMatchSnapshot();
   });
 
+  it("should render the `Stepper` with all step labels being displayed, if the right flag is provided ", async () => {
+    const props = { ...baseProps, showStepLabelWhenInactive: true, steps: 5 };
+    const { container } = render(Stepper, { ...baseOptions, props });
+
+    expect(container.firstChild).toMatchSnapshot();
+  });
+
   it("should pass additional class names and attributes to the rendered element", () => {
     const props = {
       ...baseProps,