+++ web-wallet/CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix wallet not waiting for the Wasm to load at startup [#3238]
 - Fix Public address icon in Receive flow [#3259]
 - Fix Public account icon hover text (Balance component) [#3259]
+- Fix labels in Stepper not displaying properly [#3265]
 
 ## [0.10.1] - 2024-12-18
 
@@ -478,6 +479,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3238]: https://github.com/dusk-network/rusk/issues/3238
 [#3245]: https://github.com/dusk-network/rusk/issues/3245
 [#3259]: https://github.com/dusk-network/rusk/issues/3259
+[#3265]: https://github.com/dusk-network/rusk/issues/3265
 
 <!-- VERSIONS -->